Created books reducer using createSlice
